+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Calling `ship-mate-edit-environment` from outside a `ship-mate`
   buffer now completes `ship-mate` buffers in the project to edit
   their environment.
+- Calling `ship-mate-command` with numeric prefix 5 now prompts the
+  user to edit the environment in the minibuffer.
 
 ### Changed